Move resource specific utility modules to resource utils directory
