+++ csunplugged/resources/management/commands/makeresources.py
@@ -8,8 +8,8 @@ from django.core.management.base import BaseCommand
 from django.http.request import QueryDict
 from resources.models import Resource
 from resources.views.views import generate_resource_pdf
-from utils.get_resource_generator import get_resource_generator
-from utils.resource_valid_test_configurations import resource_valid_test_configurations
+from resources.utils.get_resource_generator import get_resource_generator
+from resources.utils.resource_valid_test_configurations import resource_valid_test_configurations
 
 
 class Command(BaseCommand):
