+++ csunplugged/tests/resources/views/test_sorting_network_cards.py
@@ -3,9 +3,9 @@ from django.test import tag
 from django.urls import reverse
 from tests.BaseTestWithDB import BaseTestWithDB
 from tests.resources.ResourcesTestDataGenerator import ResourcesTestDataGenerator
-from utils.get_resource_generator import get_resource_generator
+from resources.utils.get_resource_generator import get_resource_generator
 from utils.create_query_string import query_string
-from utils.resource_valid_test_configurations import resource_valid_test_configurations
+from resources.utils.resource_valid_test_configurations import resource_valid_test_configurations
 
 
 @tag('resource_generation')
